requirements.txt
streamlit
google-api-python-client
plotly
pandas
numpy
requests
pillow
edge-tts
moviepy
groq
piper-tts
piper
ffmpeg-python
